config: interpret more inputs as unquoted strings
Currently, typing names with symbols that are part of TOML syntax can be
difficult:

```shell
# Doesn't work
$ jj config set user.name "Michael D'Angelo"

# Required to write it as a TOML string
$ jj config set user.name "'''Michael D'Angelo'''"
```

The user should know their own shell's quoting rules, so allowing
unquoted strings to only include "bare characters" (as defined by
`is_bare_char`) is unnecessary.

Relaxing this check means the user gets what they probably expect. The
following are all treated as the TOML value they look like:

| Input     | TOML value  |
|-----------|-------------|
| `foo`     | `"foo"`     |
| `123`     | `123`       |
| `12q3`    | `"12q3"`    |
| `[1,2]`   | `[1, 2]`    |
| `{a = 1}` | `{ a = 1 }` |

If the input begins with a character that indicates the user was
probably trying to input a TOML value, then we do not fall back to
interpreting it as an unquoted string:

- `'foo`
- `"foo`
- `[1,2`
- `{b:4}` (TOML uses equal sign, not colon)

Fixes #5748